Add `Backup`, `Stats`, and `Vault` commands to CLI
- Introduced `Backup` command for managing backups, including configuration, creation, listing, restoration, cleanup, and verification.
- Added `Stats` command to display vault statistics, such as item counts, password strength analysis, activity trends, and recommendations.
- Implemented `Vault` command for managing vaults, including listing, creation, switching, and deletion functionality.
- Enhanced CLI capabilities to streamline user operations and data management.
