+++ crates/cli/src/lib.rs
@@ -1,876 +1,25 @@
+mod backup;
 mod health;
+mod stats;
 mod utils;
+mod vault;
 
+use crate::backup::{BackupCommand, handle_backup_command};
 use crate::health::{analyze_password_strength, handle_health_command};
+use crate::stats::handle_stats_command;
 use crate::utils::{filter_and_sort_items, format_relative_time};
-use chamber_backup::BackupManager;
+use crate::vault::{VaultCommand, handle_vault_command};
 use chamber_import_export::{ExportFormat, detect_format_from_extension, export_items, import_items};
 use chamber_password_gen::{
     PasswordConfig, generate_complex_password, generate_memorable_password, generate_simple_password,
 };
-use chamber_vault::{BackupConfig, Item, ItemKind, NewItem, Vault, VaultCategory, VaultManager};
+use chamber_vault::{Item, ItemKind, NewItem, Vault, VaultManager};
 use clap::{Parser, Subcommand};
 use color_eyre::Result;
 use color_eyre::eyre::eyre;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::str::FromStr;
-use time::{Duration, OffsetDateTime};
-
-#[derive(Subcommand, Debug)]
-pub enum VaultCommand {
-    /// List all available vaults with their status and information
-    List,
-    /// Create a new vault with separate encryption and organization
-    Create {
-        /// Name for the new vault (e.g., "work-secrets", "personal")
-        name: String,
-        /// Custom file path for vault storage (optional)
-        #[arg(long)]
-        path: Option<PathBuf>,
-        /// Vault category for organization: personal, work, team, project, testing, archive
-        #[arg(long, default_value = "personal")]
-        category: String,
-        /// Optional description explaining the vault's purpose
-        #[arg(long)]
-        description: Option<String>,
-    },
-    /// Switch to a different vault (makes it the active vault for operations)
-    Switch {
-        /// Vault ID or name to switch to
-        vault: String,
-    },
-    /// Show information about the currently active vault
-    Active,
-    /// Delete a vault and optionally remove its database file
-    Delete {
-        /// Vault ID or name to delete
-        vault: String,
-        /// Also delete the vault database file from disk (DESTRUCTIVE)
-        #[arg(long)]
-        delete_file: bool,
-    },
-    /// Import an existing vault database file into the vault registry
-    Import {
-        file: PathBuf,
-        name: String,
-        #[arg(long, default_value = "personal")]
-        category: String,
-        #[arg(long)]
-        copy: bool,
-    },
-    /// Show vault information
-    Info { vault: Option<String> },
-    /// Update vault metadata
-    Update {
-        /// Path to the vault database file to import
-        vault: String,
-        /// Display name for the imported vault
-        #[arg(long)]
-        name: Option<String>,
-        #[arg(long)]
-        description: Option<String>,
-        /// Category for organizing the imported vault
-        #[arg(long)]
-        category: Option<String>,
-        #[arg(long)]
-        favorite: Option<bool>,
-    },
-}
-
-/// Handles various vault-related commands by interacting with a `VaultManager` instance.
-///
-/// # Arguments
-/// - `manager`: A mutable reference to the `VaultManager` instance responsible for managing vaults.
-/// - `cmd`: A `VaultCommand` specifying the action to be performed (e.g., listing vaults, creating a vault).
-///
-/// # Returns
-/// - `Result<()>`: Returns `Ok(())` if the command was processed successfully, or an error if something went wrong.
-///
-/// # Supported Commands
-///
-/// 1. **List (`VaultCommand::List`)**:
-///    Lists all the vaults managed by the `VaultManager`. Each vault displays its ID, name, category, path, creation time,
-///    description (if available), active status, and favorite status.
-///
-/// 2. **Create (`VaultCommand::Create`)**:
-///    Creates a new vault with the specified name, path, category, and optional description.
-///    It prompts the user for a master password and generates a new vault with a unique ID.
-///
-///    - `name`: The name of the new vault.
-///    - `path`: The directory where the vault will be stored.
-///    - `category`: A category to organize the vault.
-///    - `description`: (Optional) Additional details about the vault.
-///
-/// 3. **Switch (`VaultCommand::Switch`)**:
-///    Switches the active vault to the one specified by its ID or name. Updates the `VaultManager` to reflect this change.
-///
-/// 4. **Active (`VaultCommand::Active`)**:
-///    Displays information about the currently active vault, including its name, ID, path, category, and description
-///    (if set). If no vault is active, it notifies the user.
-///
-/// 5. **Delete (`VaultCommand::Delete`)**:
-///    Deletes a specified vault either by its ID or name. If `delete_file` is true, the associated vault file is also
-///    removed. The user is prompted for confirmation before deletion.
-///
-///    - `vault`: The ID or name of the vault to be deleted.
-///    - `delete_file`: A flag indicating whether the file on disk should also be deleted.
-///
-/// 6. **Import (`VaultCommand::Import`)**:
-///    Imports an existing vault from a specified file. The user can assign a new name and category, and optionally copy
-///    the vault file to the vault directory.
-///
-///    - `file`: The path to the file to import.
-///    - `name`: The name of the imported vault.
-///    - `category`: A category to assign to the newly imported vault.
-///    - `copy`: A flag indicating whether to copy the file or reference it directly.
-///
-/// 7. **Info (`VaultCommand::Info`)**:
-///    Retrieves and displays detailed information about a vault. If no specific vault is specified, it uses the currently
-///    active vault.
-///
-///    - `vault`: Optionally specify a vault by ID or name. If not provided, the active vault is used.
-///
-/// 8. **Update (`VaultCommand::Update`)**:
-///    Updates the details of a specified vault, including name, description, category, and favorite status.
-///
-///    - `vault`: The ID or name of the vault to update.
-///    - `name`: (Optional) A new name for the vault.
-///    - `description`: (Optional) A new description for the vault.
-///    - `category`: (Optional) A new category for the vault. Will be parsed to validate.
-///    - `favorite`: (Optional) Updates the favorite status.
-///
-/// # Errors
-/// This function may return an error in the following cases:
-/// - Invalid input or data (e.g., an invalid category or vault ID).
-/// - Operational issues such as file read/write errors, or issues updating the `VaultManager`.
-/// - Failure to interact with the user for input or confirmation.
-#[allow(clippy::too_many_lines)]
-pub fn handle_vault_command(manager: &mut VaultManager, cmd: VaultCommand) -> Result<()> {
-    match cmd {
-        VaultCommand::List => {
-            let manager = VaultManager::new()?;
-            let vaults = manager.list_vaults();
-
-            if vaults.is_empty() {
-                println!("No vaults found. Create your first vault with:");
-                println!("  chamber registry create <name>");
-                return Ok(());
-            }
-
-            println!("📁 Available Vaults:");
-            println!("━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━");
-
-            for vault in vaults {
-                let status = if manager.is_vault_open(&vault.id) {
-                    "Open"
-                } else {
-                    "Closed"
-                };
-                let active_indicator = if Some(&vault.id) == manager.registry.active_vault_id.as_ref() {
-                    " ← Active"
-                } else {
-                    ""
-                };
-
-                println!("ID: {}           Name: {}{}", vault.id, vault.name, active_indicator);
-                println!("Category: {}          Status: {}", vault.category, status);
-                if let Some(desc) = &vault.description {
-                    println!("Description: {desc}");
-                }
-                println!("Path: {}", vault.path.display());
-                println!("━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━");
-            }
-        }
-
-        VaultCommand::Create {
-            name,
-            path,
-            category,
-            description,
-        } => {
-            let category = parse_category(&category);
-            let password = rpassword::prompt_password("Enter master password for new vault: ")?;
-            let vault_id = manager.create_vault(name.clone(), path, category, description, &password)?;
-            println!("Created vault '{name}' with ID: {vault_id}");
-        }
-
-        VaultCommand::Switch { vault } => {
-            // Try to find vault by ID or name
-            let vault_id = find_vault_id(manager, &vault)?;
-            manager.switch_active_vault(&vault_id)?;
-            println!("Switched to vault: {vault_id}");
-        }
-
-        VaultCommand::Active => {
-            if let Some(active) = manager.registry.get_active_vault() {
-                println!("Active vault: {} ({})", active.name, active.id);
-                println!("  Path: {}", active.path.display());
-                println!("  Category: {}", active.category);
-                if let Some(desc) = &active.description {
-                    println!("  Description: {desc}");
-                }
-            } else {
-                println!("No active vault");
-            }
-        }
-
-        VaultCommand::Delete { vault, delete_file } => {
-            let vault_id = find_vault_id(manager, &vault)?;
-
-            print!("Are you sure you want to delete vault '{vault_id}'? ");
-            if delete_file {
-                print!("This will also delete the vault file! ");
-            }
-            print!("(y/N): ");
-
-            let mut input = String::new();
-            std::io::stdin().read_line(&mut input)?;
-
-            if input.trim().to_lowercase() == "y" {
-                manager.delete_vault(&vault_id, delete_file)?;
-                println!("Deleted vault: {vault_id}");
-            } else {
-                println!("Cancelled");
-            }
-        }
-
-        VaultCommand::Import {
-            file,
-            name,
-            category,
-            copy,
-        } => {
-            let category = parse_category(&category);
-            let vault_id = manager.import_vault(&file, name.clone(), category, copy)?;
-            println!("Imported vault '{name}' with ID: {vault_id}");
-        }
-
-        VaultCommand::Info { vault } => {
-            let vault_id = if let Some(v) = vault {
-                find_vault_id(manager, &v)?
-            } else {
-                manager
-                    .registry
-                    .get_active_vault()
-                    .ok_or_else(|| eyre!("No active vault"))?
-                    .id
-                    .clone()
-            };
-
-            if let Some(info) = manager.registry.get_vault(&vault_id) {
-                println!("Vault Information:");
-                println!("  ID: {}", info.id);
-                println!("  Name: {}", info.name);
-                println!("  Category: {}", info.category);
-                println!("  Path: {}", info.path.display());
-                println!(
-                    "  Created: {}",
-                    info.created_at.format(&time::format_description::well_known::Rfc3339)?
-                );
-                println!(
-                    "  Last accessed: {}",
-                    info.last_accessed
-                        .format(&time::format_description::well_known::Rfc3339)?
-                );
-                println!("  Active: {}", info.is_active);
-                println!("  Favorite: {}", info.is_favorite);
-
-                if let Some(desc) = &info.description {
-                    println!("  Description: {desc}");
-                }
-            }
-        }
-
-        VaultCommand::Update {
-            vault,
-            name,
-            description,
-            category,
-            favorite,
-        } => {
-            let vault_id = find_vault_id(manager, &vault)?;
-            let category = category.map(|cat| parse_category(&cat));
-
-            manager.update_vault_info(&vault_id, name, description, category, favorite)?;
-            println!("Updated vault: {vault_id}");
-        }
-    }
-
-    Ok(())
-}
-
-fn parse_category(category: &str) -> VaultCategory {
-    match category.to_lowercase().as_str() {
-        "personal" => VaultCategory::Personal,
-        "work" => VaultCategory::Work,
-        "team" => VaultCategory::Team,
-        "project" => VaultCategory::Project,
-        "testing" => VaultCategory::Testing,
-        "archive" => VaultCategory::Archive,
-        custom => VaultCategory::Custom(custom.to_string()),
-    }
-}
-
-fn find_vault_id(manager: &VaultManager, identifier: &str) -> Result<String> {
-    // First try exact ID match
-    if manager.registry.vaults.contains_key(identifier) {
-        return Ok(identifier.to_string());
-    }
-
-    // Then try name match
-    for (id, vault_info) in &manager.registry.vaults {
-        if vault_info.name.to_lowercase() == identifier.to_lowercase() {
-            return Ok(id.clone());
-        }
-    }
-
-    Err(eyre!("Vault '{}' not found", identifier))
-}
-
-#[derive(Subcommand, Debug)]
-pub enum BackupCommand {
-    /// Configure automatic backups
-    Configure {
-        /// Enable or disable automatic backups
-        #[arg(long)]
-        enable: Option<bool>,
-        /// Backup interval in hours
-        #[arg(long)]
-        interval: Option<u64>,
-        /// Maximum number of backups to retain
-        #[arg(long)]
-        max_backups: Option<usize>,
-        /// Backup directory path
-        #[arg(long)]
-        backup_dir: Option<PathBuf>,
-        /// Backup format (json, csv, backup)
-        #[arg(long)]
-        format: Option<String>,
-        /// Enable/disable compression
-        #[arg(long)]
-        compress: Option<bool>,
-        /// Enable/disable backup verification
-        #[arg(long)]
-        verify: Option<bool>,
-    },
-    /// Create a backup immediately
-    Now {
-        /// Optional custom backup path
-        #[arg(long)]
-        output: Option<PathBuf>,
-        /// Force backup even if one was recently created
-        #[arg(long)]
-        force: bool,
-    },
-    /// List all existing backups
-    List {
-        /// Show detailed information
-        #[arg(long, short)]
-        verbose: bool,
-    },
-    /// Restore from a backup file
-    Restore {
-        /// Path to the backup file
-        backup_path: PathBuf,
-        /// Skip confirmation prompt
-        #[arg(long, short)]
-        yes: bool,
-    },
-    /// Verify backup integrity
-    Verify {
-        /// Path to the backup file to verify
-        backup_path: PathBuf,
-    },
-    /// Show current backup configuration
-    Status,
-    /// Clean up old backups manually
-    Cleanup {
-        /// Number of backups to keep (overrides config)
-        #[arg(long)]
-        keep: Option<usize>,
-        /// Perform a dry run (show what would be deleted)
-        #[arg(long)]
-        dry_run: bool,
-    },
-}
-
-/// Handles backup-related commands issued by the user.
-///
-/// This function processes a given `BackupCommand` by delegating it to the appropriate
-/// handler based on the command variant. The function works on the provided `Vault` instance
-/// and executes the associated functionality.
-///
-/// # Arguments
-///
-/// * `vault` - A `Vault` instance representing the storage system or environment where backups
-///   are managed.
-/// * `cmd` - An instance of the `BackupCommand` enum, specifying the backup operation to be performed.
-///
-/// # Variants of `BackupCommand`
-///
-/// - `Configure`:
-///   Configures backup settings such as enabling/disabling backup, setting backup intervals,
-///   maximum number of backups to retain, backup directory, format, compression settings,
-///   and verification options.
-///
-///   * `enable` - Whether to enable or disable backups.
-///   * `interval` - The interval between automatic backups (if enabled).
-///   * `max_backups` - Maximum number of backups to retain.
-///   * `backup_dir` - Directory where backups are saved.
-///   * `format` - Format of the backup files.
-///   * `compress` - Whether to compress backups.
-///   * `verify` - Whether to verify backups after creation.
-///
-/// - `Now`:
-///   Immediately initiates a backup.
-///   * `output` - Optional output location for the backup.
-///   * `force` - Whether to force the backup operation, ignoring certain conditions.
-///
-/// - `List`:
-///   Lists existing backups.
-///   * `verbose` - If `true`, provides detailed information about each backup.
-///
-/// - `Restore`:
-///   Restores a specific backup.
-///   * `backup_path` - Path to the backup file to restore from.
-///   * `yes` - If `true`, skips confirmation prompts.
-///
-/// - `Verify`:
-///   Verifies the integrity of a specific backup file.
-///   * `backup_path` - Path to the backup file to verify.
-///
-/// - `Status`:
-///   Provides the current backup status, including information about recent backups
-///   and configuration status.
-///
-/// - `Cleanup`:
-///   Cleans up old or excess backups based on retention rules.
-///   * `keep` - Number of recent backups to retain.
-///   * `dry_run` - If `true`, performs a simulation of the cleanup process without deleting files.
-///
-/// # Returns
-///
-/// * `Ok(())` - If the specified command is successfully executed.
-/// * `Err` - If an error occurs during processing. The specific error depends on the invoked operation.
-///
-/// # Errors
-///
-/// This function will return an error if:
-/// - The supplied arguments are invalid.
-/// - The `Vault` is inaccessible or improperly configured.
-/// - File operations (e.g., reading/writing backups) encounter an issue.
-/// - A requested backup file is not found, corrupted, or fails verification.
-///
-pub fn handle_backup_command(vault: Vault, cmd: BackupCommand) -> Result<()> {
-    match cmd {
-        BackupCommand::Configure {
-            enable,
-            interval,
-            max_backups,
-            backup_dir,
-            format,
-            compress,
-            verify,
-        } => handle_configure(
-            &vault,
-            enable,
-            interval,
-            max_backups,
-            backup_dir,
-            format,
-            compress,
-            verify,
-        ),
-
-        BackupCommand::Now { output, force } => handle_backup_now(vault, output, force),
-
-        BackupCommand::List { verbose } => handle_list_backups(vault, verbose),
-
-        BackupCommand::Restore { backup_path, yes } => handle_restore_backup(vault, &backup_path, yes),
-
-        BackupCommand::Verify { backup_path } => handle_verify_backup(&backup_path),
-
-        BackupCommand::Status => handle_backup_status(vault),
-
-        BackupCommand::Cleanup { keep, dry_run } => handle_cleanup_backups(vault, keep, dry_run),
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn handle_configure(
-    vault: &Vault,
-    enable: Option<bool>,
-    interval: Option<u64>,
-    max_backups: Option<usize>,
-    backup_dir: Option<PathBuf>,
-    format: Option<String>,
-    compress: Option<bool>,
-    verify: Option<bool>,
-) -> Result<()> {
-    let mut config = vault.get_backup_config().unwrap_or_default();
-    let mut changed = false;
-
-    if let Some(enabled) = enable {
-        config.enabled = enabled;
-        changed = true;
-        println!("Automatic backups {}", if enabled { "enabled" } else { "disabled" });
-    }
-
-    if let Some(hours) = interval {
-        if hours == 0 {
-            return Err(eyre!("Backup interval must be greater than 0"));
-        }
-        config.interval_hours = hours;
-        changed = true;
-        println!("Backup interval set to {hours} hours");
-    }
-
-    if let Some(max) = max_backups {
-        if max == 0 {
-            return Err(eyre!("Max backups must be greater than 0"));
-        }
-        config.max_backups = max;
-        changed = true;
-        println!("Maximum backups set to {max}");
-    }
-
-    if let Some(dir) = backup_dir {
-        config.backup_dir.clone_from(&dir);
-        changed = true;
-        println!("Backup directory set to: {}", dir.display());
-    }
-
-    if let Some(fmt) = format {
-        match fmt.as_str() {
-            "json" | "csv" | "backup" => {
-                config.format.clone_from(&fmt);
-                changed = true;
-                println!("Backup format set to: {fmt}");
-            }
-            _ => return Err(eyre!("Invalid format '{}'. Use: json, csv, or backup", fmt)),
-        }
-    }
-
-    if let Some(comp) = compress {
-        config.compress = comp;
-        changed = true;
-        println!("Compression {}", if comp { "enabled" } else { "disabled" });
-    }
-
-    if let Some(ver) = verify {
-        config.verify_after_backup = ver;
-        changed = true;
-        println!("Backup verification {}", if ver { "enabled" } else { "disabled" });
-    }
-
-    if changed {
-        vault.set_backup_config(&config)?;
-        println!("Backup configuration updated successfully");
-    } else {
-        println!("No configuration changes specified");
-        print_backup_config(&vault.get_backup_config().unwrap_or_default());
-    }
-
-    Ok(())
-}
-
-fn handle_backup_now(mut vault: Vault, output: Option<PathBuf>, force: bool) -> Result<()> {
-    // Unlock vault if needed
-    if !vault.is_unlocked() {
-        let password = rpassword::prompt_password("Enter master password: ")?;
-        vault.unlock(&password)?;
-    }
-
-    let config = if let Some(custom_path) = output {
-        let mut config = vault.get_backup_config().unwrap_or_default();
-        config.backup_dir = custom_path
-            .parent()
-            .unwrap_or_else(|| std::path::Path::new("."))
-            .to_path_buf();
-        config
-    } else {
-        vault.get_backup_config().unwrap_or_default()
-    };
-
-    let mut backup_manager = BackupManager::new(vault, config);
-
-    println!("Creating backup...");
-    let backup_path = if force {
-        backup_manager.force_backup()?
-    } else {
-        backup_manager
-            .backup_if_needed()?
-            .ok_or_else(|| eyre!("No backup needed (use --force to create anyway)"))?
-    };
-
-    println!("✅ Backup created successfully: {}", backup_path.display());
-
-    // Show backup info
-    let metadata = std::fs::metadata(&backup_path)?;
-    println!("   Size: {} bytes", metadata.len());
-    println!(
-        "   Created: {}",
-        chrono::DateTime::<chrono::Utc>::from(metadata.created()?).format("%Y-%m-%d %H:%M:%S UTC")
-    );
-
-    Ok(())
-}
-
-fn handle_list_backups(vault: Vault, verbose: bool) -> Result<()> {
-    let config = vault.get_backup_config().unwrap_or_default();
-    let backup_manager = BackupManager::new(vault, config);
-
-    let backups = backup_manager.find_all_backups()?;
-
-    if backups.is_empty() {
-        println!("No backups found in: {}", backup_manager.config.backup_dir.display());
-        return Ok(());
-    }
-
-    println!("Found {} backup(s):", backups.len());
-    println!();
-
-    #[allow(clippy::cast_precision_loss)]
-    #[allow(clippy::expect_used)]
-    for (i, backup_path) in backups.iter().enumerate() {
-        let metadata = std::fs::metadata(backup_path)?;
-        let size = metadata.len();
-        let modified = metadata.modified()?;
-        let datetime = chrono::DateTime::<chrono::Utc>::from(modified);
-
-        println!(
-            "{}. {}",
-            i + 1,
-            backup_path
-                .file_name()
-                .expect("Unable to get the file name")
-                .to_string_lossy()
-        );
-        println!("   Path: {}", backup_path.display());
-        println!("   Size: {} bytes ({:.2} MB)", size, size as f64 / 1_048_576.0);
-        println!("   Date: {}", datetime.format("%Y-%m-%d %H:%M:%S UTC"));
-
-        if verbose {
-            // Try to extract additional info from the backup
-            if let Some(timestamp) = backup_manager.extract_timestamp_from_filename(backup_path) {
-                println!(
-                    "   Timestamp: {}",
-                    timestamp.format(&time::format_description::well_known::Rfc3339)?
-                );
-            }
-        }
-        println!();
-    }
-
-    Ok(())
-}
-
-fn handle_restore_backup(mut vault: Vault, backup_path: &Path, skip_confirmation: bool) -> Result<()> {
-    if !backup_path.exists() {
-        return Err(eyre!("Backup file not found: {}", backup_path.display()));
-    }
-
-    // Verify backup first
-    println!("Verifying backup integrity...");
-    verify_backup_file(backup_path)?;
-    println!("✅ Backup verification passed");
-
-    if !skip_confirmation {
-        println!("⚠️  WARNING: This will replace all current vault data!");
-        print!("Are you sure you want to restore from backup? (y/N): ");
-        std::io::Write::flush(&mut std::io::stdout())?;
-
-        let mut input = String::new();
-        std::io::stdin().read_line(&mut input)?;
-
-        if !matches!(input.trim().to_lowercase().as_str(), "y" | "yes") {
-            println!("Restore cancelled");
-            return Ok(());
-        }
-    }
-
-    // Unlock vault
-    let password = rpassword::prompt_password("Enter master password: ")?;
-    vault.unlock(&password)?;
-
-    // Detect backup format
-    let format = chamber_import_export::detect_format_from_extension(backup_path)
-        .unwrap_or(chamber_import_export::ExportFormat::Json);
-
-    println!("Importing backup data...");
-    let items = chamber_import_export::import_items(backup_path, &format)?;
-
-    println!("Found {} items in backup", items.len());
-
-    // Clear existing items (if user confirmed)
-    let existing_items = vault.list_items()?;
-    for item in existing_items {
-        vault.delete_item(item.id)?;
-    }
-
-    // Import new items
-    let mut imported_count = 0;
-    for item in items {
-        match vault.create_item(&item) {
-            Ok(()) => imported_count += 1,
-            Err(e) => println!("⚠️  Failed to import '{}': {}", item.name, e),
-        }
-    }
-
-    println!("✅ Successfully imported {imported_count} items from backup");
-    Ok(())
-}
-
-fn handle_verify_backup(backup_path: &PathBuf) -> Result<()> {
-    println!("Verifying backup: {}", backup_path.display());
-
-    verify_backup_file(backup_path)?;
-
-    println!("✅ Backup verification passed");
-
-    // Show additional info
-    let metadata = std::fs::metadata(backup_path)?;
-    println!("   Size: {} bytes", metadata.len());
-
-    // Try to get item count
-    if let Some(format) = chamber_import_export::detect_format_from_extension(backup_path) {
-        if let Ok(items) = chamber_import_export::import_items(backup_path, &format) {
-            println!("   Items: {}", items.len());
-        }
-    }
-
-    Ok(())
-}
-
-fn handle_backup_status(vault: Vault) -> Result<()> {
-    let config = vault.get_backup_config().unwrap_or_default();
-
-    println!("🔒 Backup Configuration Status");
-    println!("━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━");
-    print_backup_config(&config);
-
-    // Show last backup info
-    let backup_manager = BackupManager::new(vault, config);
-    if let Ok(Some(last_backup)) = backup_manager.find_most_recent_backup() {
-        println!();
-        println!("📁 Most Recent Backup:");
-        println!("   File: {}", last_backup.display());
-
-        if let Ok(metadata) = std::fs::metadata(&last_backup) {
-            let modified = metadata.modified()?;
-            let datetime = chrono::DateTime::<chrono::Utc>::from(modified);
-            println!("   Date: {}", datetime.format("%Y-%m-%d %H:%M:%S UTC"));
-            println!("   Size: {} bytes", metadata.len());
-        }
-    } else {
-        println!();
-        println!("📁 No backups found");
-    }
-
-    Ok(())
-}
-
-fn handle_cleanup_backups(vault: Vault, keep: Option<usize>, dry_run: bool) -> Result<()> {
-    let mut config = vault.get_backup_config().unwrap_or_default();
-
-    if let Some(keep_count) = keep {
-        config.max_backups = keep_count;
-    }
-
-    let backup_manager = BackupManager::new(vault, config.clone());
-    let all_backups = backup_manager.find_all_backups()?;
-
-    if all_backups.len() <= config.max_backups {
-        println!(
-            "No cleanup needed. Found {} backups, keeping {}",
-            all_backups.len(),
-            config.max_backups
-        );
-        return Ok(());
-    }
-
-    let to_delete = all_backups.len() - config.max_backups;
-    println!(
-        "Found {} backups, will {} {} oldest backup(s)",
-        all_backups.len(),
-        if dry_run { "identify" } else { "delete" },
-        to_delete
-    );
-
-    // Sort by timestamp (the oldest first for deletion)
-    let mut sorted_backups = all_backups;
-    sorted_backups.sort_by(|a, b| {
-        let time_a = backup_manager
-            .extract_timestamp_from_filename(a)
-            .unwrap_or(time::OffsetDateTime::UNIX_EPOCH);
-        let time_b = backup_manager
-            .extract_timestamp_from_filename(b)
-            .unwrap_or(time::OffsetDateTime::UNIX_EPOCH);
-        time_a.cmp(&time_b)
-    });
-
-    for (i, backup_path) in sorted_backups.iter().take(to_delete).enumerate() {
-        if dry_run {
-            println!("  [{}] Would delete: {}", i + 1, backup_path.display());
-        } else {
-            match std::fs::remove_file(backup_path) {
-                Ok(()) => println!("  [{}] Deleted: {}", i + 1, backup_path.display()),
-                Err(e) => println!("  [{}] Failed to delete {}: {}", i + 1, backup_path.display(), e),
-            }
-        }
-    }
-
-    if dry_run {
-        println!("Dry run completed. Use without --dry-run to actually delete files.");
-    } else {
-        println!("✅ Cleanup completed");
-    }
-
-    Ok(())
-}
-
-fn print_backup_config(config: &BackupConfig) {
-    println!(
-        "   Status: {}",
-        if config.enabled { "✅ Enabled" } else { "❌ Disabled" }
-    );
-    println!("   Directory: {}", config.backup_dir.display());
-    println!("   Interval: {} hours", config.interval_hours);
-    println!("   Max backups: {}", config.max_backups);
-    println!("   Format: {}", config.format);
-    println!(
-        "   Compression: {}",
-        if config.compress { "✅ Enabled" } else { "❌ Disabled" }
-    );
-    println!(
-        "   Verification: {}",
-        if config.verify_after_backup {
-            "✅ Enabled"
-        } else {
-            "❌ Disabled"
-        }
-    );
-}
-
-fn verify_backup_file(backup_path: &std::path::Path) -> Result<()> {
-    if !backup_path.exists() {
-        return Err(eyre!("Backup file does not exist"));
-    }
-
-    let metadata = std::fs::metadata(backup_path)?;
-    if metadata.len() == 0 {
-        return Err(eyre!("Backup file is empty"));
-    }
-
-    // Try to parse the backup based on format
-    if let Some(format) = detect_format_from_extension(backup_path) {
-        import_items(backup_path, &format)?;
-    } else {
-        return Err(eyre!("Unable to detect backup format"));
-    }
-
-    Ok(())
-}
+use time::OffsetDateTime;
 
 #[derive(Parser, Debug)]
 #[command(
@@ -891,7 +40,7 @@ fn verify_backup_file(backup_path: &std::path::Path) -> Result<()> {
                   1. chamber init              # Initialize your first vault\n\
                   2. chamber add -n \"github-token\" -k apikey -v \"your-token\"\n\
                   3. chamber list              # View your secrets\n\
-                  4. chamber ui                # Launch terminal interface"
+                  4. chamber                   # Launch terminal interface"
 )]
 pub struct Cli {
     #[command(subcommand)]
@@ -1015,8 +164,10 @@ pub enum Commands {
         #[arg(long)]
         skip_duplicates: bool,
     },
+    /// Show vault statistics
     Stats,
 
+    /// Update an existing secret
     Update {
         #[arg(short, long, help = "Name of the item to update")]
         name: String,
@@ -1628,269 +779,6 @@ pub fn handle_command(cmd: Commands) -> Result<()> {
     Ok(())
 }
 
-#[allow(clippy::too_many_lines)]
-fn handle_stats_command(vault: &Vault) -> Result<()> {
-    let items = vault.list_items()?;
-
-    // Get vault information if possible
-    let vault_name = match vault.get_vault_id() {
-        Ok(Some(id)) => format!(" ({id})"),
-        Ok(None) => " (Default Vault)".to_string(),
-        Err(_) => String::new(),
-    };
-
-    if items.is_empty() {
-        println!("📊 Vault Statistics{vault_name}");
-        println!("══════════════════");
-        println!("No items found in vault.");
-        return Ok(());
-    }
-
-    // Rest of the statistics code remains the same...
-    let mut type_counts = std::collections::HashMap::new();
-    let mut weak_passwords = Vec::new();
-    let mut old_items = Vec::new();
-    let mut recently_updated = Vec::new();
-    let mut duplicate_names = Vec::new();
-    let mut password_lengths = Vec::new();
-
-    let now = OffsetDateTime::now_utc();
-    let cutoff_date = now - Duration::days(180);
-    let recent_cutoff = now - Duration::days(30);
-
-    // Check for duplicate names
-    let mut name_counts = std::collections::HashMap::new();
-    for item in &items {
-        let count = name_counts.entry(&item.name).or_insert(0);
-        *count += 1;
-    }
-
-    for (name, count) in &name_counts {
-        if *count > 1 {
-            duplicate_names.push((name.as_str(), *count));
-        }
-    }
-
-    for item in &items {
-        // Count by type
-        let count = type_counts.entry(item.kind).or_insert(0);
-        *count += 1;
-
-        // Check for weak passwords and collect password lengths
-        if item.kind == chamber_vault::ItemKind::Password {
-            password_lengths.push(item.value.len());
-            if is_weak_password(&item.value) {
-                weak_passwords.push(&item.name);
-            }
-        }
-
-        // Check for old items (older than 180 days)
-        if item.created_at < cutoff_date {
-            old_items.push((&item.name, item.created_at));
-        }
-
-        // Check for recently updated items (within 30 days)
-        if item.updated_at > recent_cutoff {
-            recently_updated.push((&item.name, item.updated_at));
-        }
-    }
-
-    // Display statistics with vault name
-    println!("📊 Vault Statistics{vault_name}");
-    println!("══════════════════");
-    println!();
-
-    // Add vault location info
-    println!("📍 Vault Info:");
-    println!("──────────────");
-    println!("  📂 Database Path: {}", vault.db_path().display());
-    println!();
-
-    // General overview
-    println!("📋 Overview:");
-    println!("────────────");
-    println!("  📊 Total Items: {}", items.len());
-
-    if let Some(oldest) = items.iter().min_by_key(|item| item.created_at) {
-        let vault_age = now - oldest.created_at;
-        println!("  📅 Vault Age: {} days (oldest item)", vault_age.whole_days());
-    }
-
-    if let Some(newest) = items.iter().max_by_key(|item| item.updated_at) {
-        let last_activity = now - newest.updated_at;
-        println!("  🕒 Last Activity: {} days ago", last_activity.whole_days());
-    }
-
-    println!("  🔄 Recently Updated (30 days): {}", recently_updated.len());
-    println!();
-
-    // Items per vault type
-    println!("📁 Items by Type:");
-    println!("─────────────────");
-    let mut sorted_types: Vec<_> = type_counts.iter().collect();
-    sorted_types.sort_by(|a, b| b.1.cmp(a.1)); // Sort by count descending
-
-    #[allow(clippy::cast_precision_loss)]
-    for (item_type, count) in sorted_types {
-        let percentage = (f64::from(*count) / items.len() as f64) * 100.0;
-        println!(
-            "  {} {}: {} ({:.1}%)",
-            get_type_emoji(*item_type),
-            item_type.display_name(),
-            count,
-            percentage
-        );
-    }
-    println!();
-
-    // Password security analysis
-    if !password_lengths.is_empty() {
-        println!("🔐 Password Security:");
-        println!("────────────────────");
-
-        #[allow(clippy::cast_precision_loss)]
-        let avg_length = password_lengths.iter().sum::<usize>() as f64 / password_lengths.len() as f64;
-        let min_length = password_lengths.iter().min().unwrap_or(&0);
-        let max_length = password_lengths.iter().max().unwrap_or(&0);
-
-        println!("  📏 Average Length: {avg_length:.1} characters");
-        println!("  📐 Length Range: {min_length} - {max_length} characters");
-
-        let strong_count = password_lengths.len() - weak_passwords.len();
-        #[allow(clippy::cast_precision_loss)]
-        let strong_percentage = (strong_count as f64 / password_lengths.len() as f64) * 100.0;
-
-        println!("  💪 Strong Passwords: {strong_count} ({strong_percentage:.1}%)");
-        println!(
-            "  🔓 Weak Passwords: {} ({:.1}%)",
-            weak_passwords.len(),
-            100.0 - strong_percentage
-        );
-        println!();
-    }
-
-    // Weak passwords details
-    println!("🔓 Weak Passwords:");
-    println!("──────────────────");
-    if weak_passwords.is_empty() {
-        println!("  ✅ No weak passwords found!");
-    } else {
-        println!("  ❌ Found {} weak password(s):", weak_passwords.len());
-        for name in &weak_passwords {
-            println!("    • {name}");
-        }
-        println!("  💡 Strong passwords should have at least 10 characters with");
-        println!("     uppercase, lowercase, numbers, and special characters.");
-    }
-    println!();
-
-    // Data quality issues
-    println!("🔍 Data Quality:");
-    println!("────────────────");
-
-    // Duplicate names
-    if duplicate_names.is_empty() {
-        println!("  ✅ No duplicate names found!");
-    } else {
-        println!("  ⚠️  Found {} duplicate name(s):", duplicate_names.len());
-        duplicate_names.sort_by(|a, b| b.1.cmp(&a.1)); // Sort by count descending
-        for (name, count) in &duplicate_names {
-            println!("    • '{name}' appears {count} times");
-        }
-    }
-    println!();
-
-    // Age analysis
-    println!("⏰ Age Analysis:");
-    println!("───────────────");
-
-    let new_items = items.iter().filter(|item| item.created_at > recent_cutoff).count();
-    let medium_items = items
-        .iter()
-        .filter(|item| item.created_at <= recent_cutoff && item.created_at >= cutoff_date)
-        .count();
-
-    println!("  🆕 New Items (< 30 days): {new_items}");
-    println!("  📅 Medium Age (30-180 days): {medium_items}");
-    println!("  ⏳ Old Items (> 180 days): {}", old_items.len());
-
-    if !old_items.is_empty() {
-        println!("  ⚠️  Oldest items:");
-        let mut sorted_old_items = old_items.clone();
-        sorted_old_items.sort_by(|a, b| a.1.cmp(&b.1)); // Sort by date ascending (oldest first)
-
-        for (name, created_at) in sorted_old_items.iter().take(5) {
-            let age = now - *created_at;
-            println!("    • {} ({} days old)", name, age.whole_days());
-        }
-
-        if sorted_old_items.len() > 5 {
-            println!("    ... and {} more", sorted_old_items.len() - 5);
-        }
-
-        println!("  💡 Consider updating or reviewing old items for security.");
-    }
-
-    println!();
-
-    // Activity summary
-    if !recently_updated.is_empty() {
-        println!("🔄 Recent Activity:");
-        println!("──────────────────");
-
-        let mut sorted_recent = recently_updated;
-        sorted_recent.sort_by(|a, b| b.1.cmp(&a.1)); // Sort by date descending (most recent first)
-
-        println!("  📝 Recently updated items:");
-        for (name, updated_at) in sorted_recent.iter().take(5) {
-            let age = now - *updated_at;
-            println!("    • {} ({} days ago)", name, age.whole_days());
-        }
-
-        if sorted_recent.len() > 5 {
-            println!("    ... and {} more", sorted_recent.len() - 5);
-        }
-        println!();
-    }
-
-    // Security recommendations
-    println!("💡 Security Recommendations:");
-    println!("────────────────────────────");
-
-    let mut recommendations = Vec::new();
-
-    if !weak_passwords.is_empty() {
-        recommendations.push(format!("Update {} weak password(s)", weak_passwords.len()));
-    }
-
-    if !old_items.is_empty() {
-        recommendations.push(format!("Review {} old item(s) (>180 days)", old_items.len()));
-    }
-
-    if !duplicate_names.is_empty() {
-        recommendations.push(format!("Resolve {} duplicate name(s)", duplicate_names.len()));
-    }
-
-    let very_old_items = items
-        .iter()
-        .filter(|item| item.created_at < (now - Duration::days(365)))
-        .count();
-
-    if very_old_items > 0 {
-        recommendations.push(format!("Consider rotating {very_old_items} very old item(s) (>1 year)"));
-    }
-
-    if recommendations.is_empty() {
-        println!("  ✅ Your vault looks healthy! No immediate actions needed.");
-    } else {
-        for (i, rec) in recommendations.iter().enumerate() {
-            println!("  {}. {}", i + 1, rec);
-        }
-    }
-
-    Ok(())
-}
-
 fn is_weak_password(password: &str) -> bool {
     // Strong password criteria:
     // - At least 10 characters
@@ -1916,10 +804,19 @@ const fn get_type_emoji(item_type: ItemKind) -> &'static str {
         ItemKind::Password => "🔑",
         ItemKind::EnvVar => "🌐",
         ItemKind::Note => "📝",
-        ItemKind::ApiKey => "🔐",
-        ItemKind::SshKey => "🔒",
+        ItemKind::ApiKey | ItemKind::SshKey => "🔐",
         ItemKind::Certificate => "📜",
         ItemKind::Database => "🗄️",
+        ItemKind::CreditCard => "💳",
+        ItemKind::SecureNote => "🔒",
+        ItemKind::Identity => "🆔",
+        ItemKind::Server => "🖥️",
+        ItemKind::WifiPassword => "📶",
+        ItemKind::License => "📄",
+        ItemKind::BankAccount => "🏦",
+        ItemKind::Document => "📋",
+        ItemKind::Recovery => "🔄",
+        ItemKind::OAuth => "🎫",
     }
 }
